Move tracking of rack TabIndex into RackState
